+++ frontend/src/composables/useCmsProductionGrid.ts
@@ -1,12 +1,8 @@
-import { computed, ref, type Ref } from "vue";
+import { computed, type Ref } from "vue";
 import type {
   ColDef,
-  GridApi,
-  GridReadyEvent,
-  GridState,
   Module,
   RowClassParams,
-  SelectionChangedEvent,
 } from "ag-grid-community";
 import {
   AllCommunityModule,
@@ -55,6 +51,7 @@ import {
 } from "ag-grid-community";
 import type { ICellRendererParams } from "ag-grid-community";
 import type { CmsProductionGridRow } from "@/services/cms";
+import { useCmsGridBase } from "./useCmsGridBase";
 
 type TranslateFunction = (key: string, params?: Record<string, unknown>) => string;
 
@@ -159,33 +156,30 @@ function renderMediaCell(value: unknown): string {
 }
 
 /**
- * Encapsulates CMS AG Grid configuration, state, and utility actions.
- *
- * Responsibilities:
- * - define column config and editor behavior
- * - persist/restore grid state to localStorage
- * - expose actions for filtering, exporting, and sizing
- * - expose theme variables compatible with light/dark modes
+ * CMS productions grid: layers production-specific column defs, cell styling
+ * for empty/finalized rows, and a custom CSV cell processor on top of the
+ * shared {@link useCmsGridBase} plumbing.
  */
 export function useCmsProductionGrid(options: {
   isDark: Ref<boolean>;
   t: TranslateFunction;
   getPrimaryTagLabels?: () => string[];
 }) {
-  const gridApi = ref<GridApi<CmsProductionGridRow> | null>(null);
-  const quickFilterText = ref("");
-  const selectedCount = ref(0);
-  const columnChooserOpen = ref(false);
-  const columnVisibility = ref<Record<string, boolean>>(
-    Object.fromEntries(cmsGridColumnIds.map((colId) => [colId, true])) as Record<string, boolean>,
-  );
+  const base = useCmsGridBase<CmsProductionGridRow>({
+    isDark: options.isDark,
+    storageKey: cmsGridStateStorageKey,
+    columnIds: cmsGridColumnIds,
+    csvFileName: "cms-productions.csv",
+    csvExcludedColumnIds: ["eventsAction"],
+    csvProcessCellCallback: (params) => {
+      if (params.column.getColId() === "tags") {
+        const data = params.node?.data as CmsProductionGridRow | undefined;
+        return JSON.stringify(data?.source.tags ?? []);
+      }
 
-  const rowSelection = {
-    mode: "multiRow" as const,
-    checkboxes: true,
-    headerCheckbox: true,
-    enableClickSelection: true,
-  };
+      return String(params.value ?? "");
+    },
+  });
 
   const defaultColDef: ColDef<CmsProductionGridRow> = {
     editable: false,
@@ -308,179 +302,6 @@ export function useCmsProductionGrid(options: {
     },
   ]);
 
-  const agThemeVars = computed<Record<string, string>>(() => {
-    if (options.isDark.value) {
-      return {
-        "--ag-background-color": "var(--surface-0)",
-        "--ag-foreground-color": "var(--ink-primary)",
-        "--ag-header-background-color": "var(--surface-inv-raised)",
-        "--ag-header-foreground-color": "var(--ink-on-inv)",
-        "--ag-odd-row-background-color": "color-mix(in srgb, var(--surface-inv-raised) 55%, transparent)",
-        "--ag-row-hover-color": "color-mix(in srgb, var(--surface-inv-border) 65%, transparent)",
-        "--ag-border-color": "var(--surface-inv-border)",
-        "--ag-selected-row-background-color": "transparent",
-        "--ag-range-selection-background-color": "transparent",
-        "--ag-header-column-separator-color": "var(--surface-inv-border)",
-        "--ag-input-focus-border-color": "var(--surface-inv)",
-        "--ag-font-family": '"Inter Variable", sans-serif',
-        "--ag-font-size": "13px",
-        "--cms-selected-row-bg": "color-mix(in srgb, var(--surface-inv-raised) 42%, transparent)",
-        "--cms-header-fg": "var(--ink-on-inv)",
-        "--cms-checkbox-color": "var(--ink-on-inv)",
-      };
-    }
-
-    return {
-      "--ag-background-color": "var(--surface-0)",
-      "--ag-foreground-color": "var(--ink-primary)",
-      "--ag-header-background-color": "var(--surface-1)",
-      "--ag-header-foreground-color": "var(--ink-primary)",
-      "--ag-odd-row-background-color": "var(--surface-1)",
-      "--ag-row-hover-color": "color-mix(in srgb, var(--surface-2) 70%, transparent)",
-      "--ag-border-color": "var(--surface-3)",
-      "--ag-selected-row-background-color": "transparent",
-      "--ag-range-selection-background-color": "transparent",
-      "--ag-header-column-separator-color": "var(--surface-3)",
-      "--ag-input-focus-border-color": "var(--surface-inv)",
-      "--ag-font-family": '"Inter Variable", sans-serif',
-      "--ag-font-size": "13px",
-      "--cms-selected-row-bg": "color-mix(in srgb, var(--surface-2) 55%, transparent)",
-      "--cms-header-fg": "var(--ink-primary)",
-      "--cms-checkbox-color": "var(--ink-primary)",
-    };
-  });
-
-  function loadPersistedGridState(): GridState | null {
-    const rawState = localStorage.getItem(cmsGridStateStorageKey);
-    if (!rawState) {
-      return null;
-    }
-
-    try {
-      return JSON.parse(rawState) as GridState;
-    } catch {
-      return null;
-    }
-  }
-
-  function persistGridState(): void {
-    if (!gridApi.value) {
-      return;
-    }
-
-    localStorage.setItem(cmsGridStateStorageKey, JSON.stringify(gridApi.value.getState()));
-  }
-
-  function syncColumnVisibilityFromGrid(): void {
-    const state = gridApi.value?.getColumnState() ?? [];
-    const nextVisibility: Record<string, boolean> = { ...columnVisibility.value };
-
-    for (const column of state) {
-      if (column.colId) {
-        nextVisibility[column.colId] = column.hide !== true;
-      }
-    }
-
-    columnVisibility.value = nextVisibility;
-  }
-
-  function restoreGridState(): boolean {
-    const state = loadPersistedGridState();
-    if (!gridApi.value || !state) {
-      return false;
-    }
-
-    gridApi.value.setState(state);
-    syncColumnVisibilityFromGrid();
-    return true;
-  }
-
-  function setGridColumnVisibility(colId: string, visible: boolean): void {
-    columnVisibility.value = {
-      ...columnVisibility.value,
-      [colId]: visible,
-    };
-
-    gridApi.value?.applyColumnState({
-      state: [{ colId, hide: !visible }],
-    });
-    gridApi.value?.sizeColumnsToFit();
-    persistGridState();
-  }
-
-  function fitGridColumns(): void {
-    gridApi.value?.sizeColumnsToFit();
-    persistGridState();
-  }
-
-  function autoSizeGridColumns(): void {
-    gridApi.value?.autoSizeAllColumns();
-    persistGridState();
-  }
-
-  function resetGridFilters(): void {
-    quickFilterText.value = "";
-    gridApi.value?.setFilterModel(null);
-    applyQuickFilter();
-    persistGridState();
-  }
-
-  function exportGridCsv(): void {
-    const visibleColumnKeys = cmsGridColumnIds.filter(
-      (colId) => colId !== "eventsAction" && columnVisibility.value[colId] !== false,
-    );
-
-    gridApi.value?.exportDataAsCsv({
-      fileName: "cms-productions.csv",
-      columnKeys: visibleColumnKeys,
-      processCellCallback: (params) => {
-        if (params.column.getColId() === "tags") {
-          const data = params.node?.data as CmsProductionGridRow | undefined;
-          return JSON.stringify(data?.source.tags ?? []);
-        }
-
-        return String(params.value ?? "");
-      },
-    });
-  }
-
-  function resetGridState(): void {
-    localStorage.removeItem(cmsGridStateStorageKey);
-    quickFilterText.value = "";
-    selectedCount.value = 0;
-    columnChooserOpen.value = false;
-    gridApi.value?.deselectAll();
-    gridApi.value?.setFilterModel(null);
-    gridApi.value?.applyColumnState({ defaultState: { sort: null } });
-    gridApi.value?.applyColumnState({ defaultState: { hide: false } });
-    columnVisibility.value = Object.fromEntries(cmsGridColumnIds.map((colId) => [colId, true])) as Record<string, boolean>;
-    gridApi.value?.sizeColumnsToFit();
-    applyQuickFilter();
-    persistGridState();
-  }
-
-  function onGridReady(event: GridReadyEvent<CmsProductionGridRow>): void {
-    gridApi.value = event.api;
-
-    const restored = restoreGridState();
-    if (!restored) {
-      event.api.sizeColumnsToFit();
-    }
-
-    syncColumnVisibilityFromGrid();
-    applyQuickFilter();
-  }
-
-  function applyQuickFilter(): void {
-    gridApi.value?.setGridOption("quickFilterText", quickFilterText.value);
-    persistGridState();
-  }
-
-  function onSelectionChanged(event: SelectionChangedEvent<CmsProductionGridRow>): void {
-    selectedCount.value = event.api.getSelectedRows().length;
-    persistGridState();
-  }
-
   function getProductionRowStyle(
     params: RowClassParams<CmsProductionGridRow>,
   ): Record<string, string> | undefined {
@@ -494,27 +315,10 @@ export function useCmsProductionGrid(options: {
   }
 
   return {
-    agThemeVars,
-    autoSizeGridColumns,
-    columnChooserOpen,
+    ...base,
     columnDefs,
-    columnVisibility,
     defaultColDef,
-    exportGridCsv,
-    fitGridColumns,
     getProductionRowStyle,
     gridColumnOptions,
-    onGridReady,
-    onSelectionChanged,
-    quickFilterText,
-    resetGridFilters,
-    resetGridState,
-    rowSelection,
-    selectedCount,
-    setGridColumnVisibility,
-    applyQuickFilter,
-    gridApi,
-    persistGridState,
-    syncColumnVisibilityFromGrid,
   };
 }
